+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ### version 0.6.7, 2018/05/12
   * supports the following taggers: TaKIPI (for Polish), Alpino (Dutch)
-  * the Imposters method reimplemented
+  * the Imposters method reimplemented, via the new function imposters()
 
 
 ### version 0.6.6, 2018/04/13